+++ app/[userId]/repo-list/page.tsx
@@ -1,35 +1,8 @@
-"use client";
-
-import { ShortenText } from "@/utils/shorten-text";
 import TabTitle from "../_components/tab-title";
 import { TAB_LIST } from "@/constants/user/nav-tab";
-import { useEffect, useState } from "react";
-import { useUserStore } from "@/libs/zustand/user";
-import { getMyRepositoryList } from "@/apis/repository";
-import { MyRepositoryListType } from "@/type/user";
+import FetchRepositoryList from "./_components/fetch-repo-list";
 
 const RepositoryList = () => {
-    const [repositories, setRepositories] = useState<MyRepositoryListType[]>(
-        []
-    );
-    const { login } = useUserStore();
-
-    useEffect(() => {
-        const fetchMyRepositories = async () => {
-            if (login) {
-                try {
-                    const data = await getMyRepositoryList(login);
-                    setRepositories(data);
-                } catch {
-                    throw new Error(
-                        "회원의 repository 목록을 가져오는데 실패했어요...🫠"
-                    );
-                }
-            }
-        };
-        fetchMyRepositories();
-    }, [login]);
-
     return (
         <div className="p-10">
             <TabTitle tabNum={TAB_LIST[0]} />
@@ -41,18 +14,7 @@ const RepositoryList = () => {
                     <p className="flex-[1.5]">할당자</p>
                 </li>
                 {/* table body(list) */}
-                {repositories.map((repo, index) => (
-                    <li
-                        key={index}
-                        className="w-full h-[68px] flex-items bg-SYSTEM-white rounded-[12px] px-6 my-3 text-SYSTEM-black"
-                    >
-                        <p className="flex-[3]">
-                            {ShortenText(repo.fullName, 50)}
-                        </p>
-                        <p className="flex-[1.5]">{repo.ownerLogin}</p>
-                        {/* <p className="flex-[1.5]">{repo.assignee}</p> */}
-                    </li>
-                ))}
+                <FetchRepositoryList />
             </ul>
         </div>
     );
