--- app/_components/type.ts
@@ -1,3 +1,0 @@
-export type UserDashboardProps = {
-    userId: string;
-};
